iio: adrv9002: do not allow SSI long strobe
With the new TES version, the user might choose between short and long
strobes on the interface. The effort of supporting this on hdl would
be to high without bringing any benefit. Hence, we decided (at least, for
now) to not support this, which means that we must properly invalidate
profiles that set this...

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 3a70b9fc128b9828485a8c692a23de92a500110b)